Extract state and layout logic from App component
- Extract AuthContext for user auth and initialization
- Extract Navigation types to shared file
- Extract LoginScreen, Header, and MobileNav components
- Resolve strict mode TypeScript type import issues
- Satisfy fast-refresh ESLint rules and fix React hooks usage
