+++ app/src/types/navigation.ts
@@ -1,0 +1,1 @@
+export type Screen = 'home' | 'checkin' | 'goals' | 'constraints' | 'preferences' | 'data';
